Course Timetabling: Class Duration Types
- Added ability to define different ways how class duration is entered and used in the system (to extend the formal Minutes per Week selection).
- There is a new page (Administration > Other > Class Duration Types), where the class duration types can be managed.
  - Permission ClassDurationTypes is needed to access the page, ClassDurationTypeEdit is needed to make changes.
- Default duration type can be defined on an academic session, it can be overriden on the instructional offering configuration level (Instructional Offering Configuration page).
  - Name of the default duration type is used in the column headers (Classes, Instructional Offerings), abbreviation is used to suffix durations that are using a non-default duration type.
- Following class duration types have been implemented:
  - Minutes per Week (default): this is the existing behaviour (minutes per week == number of meetings x minutes per meeting)
  - Average Weekly Mintes: average number of minutes a class takes during a week (weekly minutes = number of meetings x minutes per meeting x number of weeks of the data pattern / number of weeks in the term)
    - Paramter: number of weeks in the term (defaults to the number of weeks of the default date pattern)
  - Semester Minutes: number of minutes in the semester (semester minutes == number of meetings x minutes per meeting x number of weeks of the data pattern)
  - Semester Hours: number of hours in the semester (semester hours == number of meetings x minutes per meeting x number of weeks of the data pattern / minutes in an hour)
    - Paramter: number of minutes in an hour (defaults to 50)
  - Meetings Minutes: number of minutes of all the meetings, excluding vacations (meeting minutes == number of meetings based on the date pattern x minutes per meeting)
    - Parameters: allowed variation of the meeting minutes, expressed as lower and upper bound (defaults to 0.95,1.10 -- meaning a class can have 5% less time, combinations with 10% more time will be also considered).
    - If the number of meetings based on the date pattern exceeds the desired number of meeting minutes, only meetings covering the given number of meeting minutes are created.
  - Meeting Hours: just like meeting minutes, but with hours (semester hours == number of meetings based on the date pattern x minutes per meeting / minutes in an hour)
    - Paramters: number of minutes in an hour and the allowed variation of the meeting minutes (defaults to 50,0.95,1.10 -- 50 minutes in an hour, 5% less time allowed, combinations with up to 10% excess time are considered).
- It is possible to hide class duration types that are not to be used by the users.
  - If only one type is visible, there is no selection of the class duration type on the Instructional Offering Configuration page.

+++ JavaSource/org/unitime/timetable/model/base/BaseDatePattern.java
@@ -40,7 +40,7 @@ public abstract class BaseDatePattern implements Serializable {
 	private Integer iOffset;
 	private Integer iType;
 	private Boolean iVisible;
-	private Integer iNumberOfWeeks;
+	private Float iNumberOfWeeks;
 
 	private Session iSession;
 	private Set<DatePattern> iParents;
@@ -84,8 +84,8 @@ public abstract class BaseDatePattern implements Serializable {
 	public Boolean getVisible() { return iVisible; }
 	public void setVisible(Boolean visible) { iVisible = visible; }
 
-	public Integer getNumberOfWeeks() { return iNumberOfWeeks; }
-	public void setNumberOfWeeks(Integer numberOfWeeks) { iNumberOfWeeks = numberOfWeeks; }
+	public Float getNumberOfWeeks() { return iNumberOfWeeks; }
+	public void setNumberOfWeeks(Float numberOfWeeks) { iNumberOfWeeks = numberOfWeeks; }
 
 	public Session getSession() { return iSession; }
 	public void setSession(Session session) { iSession = session; }
